samplers module; add QO on BaseGeometric
