+++ python/example/test.py
@@ -23,14 +23,14 @@ def main():
     select_1 = odp.trans.make_select_column(b"<HammingDistance, i32, i32>", opendp.i32_p(1))
     clamp_1 = odp.trans.make_clamp(b"<HammingDistance, i32>", opendp.i32_p(0), opendp.i32_p(10))
     bounded_sum_1 = odp.trans.make_bounded_sum(b"<HammingDistance, L1Sensitivity<i32>, i32>", opendp.i32_p(0), opendp.i32_p(10))
-    base_geometric_1 = odp.meas.make_base_simple_geometric(b"<i32>", 1.0, opendp.u32_p(0), opendp.u32_p(1000))
+    base_geometric_1 = odp.meas.make_base_simple_geometric(b"<i32, f64>", opendp.f64_p(1.0), opendp.u32_p(0), opendp.u32_p(1000))
     # base_laplace_1 = odp.meas.make_base_laplace(b"<i32>", opendp.f64_p(1.0))
     noisy_sum_1 = odp.core.make_chain_mt(base_geometric_1, odp.make_chain_tt_multi(bounded_sum_1, clamp_1, select_1))
 
     # Count, col 1
     select_2 = odp.trans.make_select_column(b"<HammingDistance, i32, f64>", opendp.i32_p(2))
     count_2 = odp.trans.make_count(b"<HammingDistance, L1Sensitivity<u32>, f64>")
-    base_geometric_2 = odp.meas.make_base_simple_geometric(b"<u32>", 1.0, opendp.u32_p(0), opendp.u32_p(1000))
+    base_geometric_2 = odp.meas.make_base_simple_geometric(b"<u32, f64>", opendp.f64_p(1.0), opendp.u32_p(0), opendp.u32_p(1000))
     noisy_count_2 = odp.core.make_chain_mt(base_geometric_2, odp.make_chain_tt_multi(count_2, select_2))
 
     # Compose & chain
