+++ rust/opendp/src/lib.rs
@@ -140,3 +140,4 @@ pub mod dom;
 pub mod meas;
 pub mod trans;
 pub mod traits;
+pub mod samplers;
